Fix display of fibers ending before a suspend
When switching to a new fiber, we record a pause event on the old one.
If the old fiber had already ended, this had the effect of extending its
lifetime to the time of the switch. Normally this doesn't matter because
we switch to the next fiber immediately, but it's a problem if there is
no runnable thread and the domain sleeps. Then the fiber appears to
remain until the domain is resumed.